// web/src/lib/generated/BucklingModeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BucklingModeDto = { eigenvalue: number, 
/**
 * Modal amplitudes a_ij (m rows of n). Feed these back into
 * [`compute_buckling_surface`] to get a plottable displacement field.
 */
shape: Array<Array<number>>, };